+++ backend/src/app.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { ConfigModule } from '@nestjs/config';
+import { ActivitiesModule } from './activities/activities.module';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { StravaModule } from './strava/strava.module';
 
 @Module({
-  imports: [ConfigModule.forRoot({ isGlobal: true }), StravaModule],
+  imports: [ConfigModule.forRoot({ isGlobal: true }), ActivitiesModule],
   controllers: [AppController],
   providers: [AppService]
 })